Prompt preflight checks that have host collectors
