+++ internal/util/util.go
@@ -1,9 +1,12 @@
 package util
 
 import (
+	"bufio"
 	"bytes"
+	"fmt"
 	"net/url"
 	"os"
+	"os/user"
 	"strings"
 	"text/template"
 
@@ -104,4 +107,39 @@ func RenderTemplate(tpl string, data interface{}) (string, error) {
 
 	// Return the string representation of the buffer
 	return buf.String(), nil
+}
+
+func IsRunningAsRoot() bool {
+	currentUser, err := user.Current()
+	if err != nil {
+		return false
+	}
+
+	// Check if the user ID is 0 (root's UID)
+	return currentUser.Uid == "0"
+}
+
+func PromptYesNo(question string) bool {
+	reader := bufio.NewReader(os.Stdin)
+
+	for {
+		fmt.Printf("%s (yes/no): ", question)
+
+		response, err := reader.ReadString('\n')
+		if err != nil {
+			fmt.Println("Error reading response:", err)
+			continue
+		}
+
+		response = strings.TrimSpace(response)
+		response = strings.ToLower(response)
+
+		if response == "yes" || response == "y" {
+			return true
+		} else if response == "no" || response == "n" {
+			return false
+		} else {
+			fmt.Println("Please type 'yes' or 'no'.")
+		}
+	}
 }
